requirements.txt
Akhet==1.0.2
Chameleon==2.5.2
Cheetah==2.2.1
ClientForm==0.2.10
Fabric==1.2.2
Flask==0.8
Jinja2==2.6
Mako==0.4.2
MarkupSafe==0.15
Paste==1.7.5.1
PasteDeploy==1.5.0
PasteScript==1.7.4.2
Pillow==1.7.5
PyStemmer==1.2.0
Pygments==1.4
Sphinx==1.1.2
Tempita==0.5.1
WebError==0.10.3
WebHelpers==1.3
WebOb==1.1beta1
WebTest==1.3.3
Werkzeug==0.8.1
ZopeSkel==2.19
cornice==0.10
coverage==3.5.1
distribute==0.6.24
docutils==0.8.1
fancycompleter==0.2
githooks==0.4dev
guess-language==0.2
hghooks==0.5.1
lxml==2.3
mechanize==0.2.5
meld3==0.6.7
ordereddict==1.1
paramiko==1.7.7.1
pdbpp==0.7.2
pycrypto==2.3
pyflakes==0.5.0
pyjslint==0.3.2
pymongo==2.0.1
pyramid==1.2.1
pyramid-mongodb==1.0
pyrepl==0.8.2
python-magic==0.4.0
pytz==2011h
repoze.lru==0.3
requests==0.8.2
simplejson==2.1.6
supervisor==3.0a10
translationstring==0.3
unittest2==0.5.1
venusian==1.0a2
virtualenv==1.6.1
virtualenvwrapper==2.7.1
virtualenvwrapper.project==1.3
wmctrl==0.1
wsgiref==0.1.2
zc.testbrowser==1.0.0a5
zest.releaser==3.22
zope.component==3.10.0
zope.configuration==3.7.4
zope.deprecation==3.4.1
zope.event==3.5.1
zope.i18nmessageid==3.6.1
zope.interface==3.7.0
zope.schema==3.8.0
zope.testbrowser==4.0.2
